refactor: separate markdown segmentation concerns
